refactor: extract services and constants
