+++ backend/src/config/constants.js
@@ -1,0 +1,3 @@
+module.exports = {
+  START_THRESHOLD: 20
+};
